requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
APScheduler==3.8.1
django-environ==0.8.1
drf-nested-routers==0.93.4
